pyproject.toml
[build-system]
requires = [ "poetry==1.0.5",]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
skip-string-normalization = true
target-version = [ "py37",]

[tool.poetry]
name = "elektrum"
version = "0.53.dev0"
description = "Media archiving & management system."
authors = [ "Edward Q. Bridges <ebridges@roja.cc>",]
readme = "README.md"
homepage = "https://elektrum.photos"
repository = "https://bitbucket.org/roja-cc/elektrum"
license = "CC-BY-NC-ND-4.0"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0"
django-allauth = "^0.41"
gunicorn = "^19.9.0"
psycopg2-binary = "^2.8.3"
boto3 = "^1.9.228"
django-sslserver = "^0.22"
djangorestframework = "^3.11"
toml = "^0.10.0"
python-dotenv = "^0.10.3"
pillow = "^7.1.1"
django-amazon-ses = "^3.0.2"
apig_wsgi = "^2.6.0"
sentry-sdk = "^0.14.4"
django-request-id = "^1.0.0"
awscli = "^1.18.73"

[tool.poetry.dev-dependencies]
pylint = "^2.3.1"
ansible-vault = "^1.2.0"
botocore = "^1.12.228"
boto = "^2.49.0"
selenium = "^3.141.0"
coverage = "^4.5.4"
factory-boy = "^2.12.0"
pytest = "5.1.2"
pytest-django = "^3.9"
pytest-cov = "^2.8"
pytest-factoryboy = "^2.0.3"
records = "^0.5.3"
virtualenvwrapper = "^4.8.4"
assertpy = "^0.14.0"
jinja2 = "^2.10.1"
cryptography = "^2.7"
jmespath-terminal = "^0.2.1"
autopep8 = "^1.4.4"
logging_tree = "^1.8"
lgw = "^1.0"
pre-commit = "^1.20"
django-extensions = "^2.2.9"
Werkzeug = "^1.0.1"
doit = "^0.32.0"

[tool.poetry.dependencies.django-storages]
git = "https://github.com/ebridges/django-storages.git"

[tool.poetry.dev-dependencies.black]
version = "^19.3b0"
allow-prereleases = true

[tool.poetry.dev-dependencies.ansible]
git = "https://github.com/ansible/ansible.git"
rev = "d77697e"
